Reorganize modules a bit
