+++ lib/core/src/text/buffer.rs
@@ -1,0 +1,1 @@
+pub mod glyph_square;
