+++ lib/core/src/text/buffer/glyph_square.rs
